Move command to DockerCredentialHelper folder
